pyproject.toml
[tool.poetry]
name = "tododdd"
version = "0.1.0"
description = ""
authors = ["Sergio Gómez <decano@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.0"
strawberry-graphql = "^0.117.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
black = {version = "^22.6.0", allow-prereleases = true}
mypy = "^0.961"
pytest-asyncio = "^0.18.3"
pytest-django = "^4.5.2"
pytest-sugar = "^0.9.4"
pytest-mock = "^3.7.0"
debugpy = "^1.6.0"
django-stubs = "^1.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
